Fix byteswap on LP64 (amd64)
kern/linux/linux_system.cc:
    Formatting
sim/host.hh:
    When we say ULL and UL we really just want a 64-bit value, not
    necessarily a long long, so cast the integer to (u)int64_t
    this fixes a problem with byte swapping.

--HG--
extra : convert_revision : ad25ace5a7b92a7c928f2d52e531193c91f3f8c8

+++ sim/host.hh
@@ -38,9 +38,9 @@
 #include <inttypes.h>
 
 /** uint64_t constant */
-#define ULL(N)		N##ULL
+#define ULL(N)		((uint64_t)N##ULL)
 /** int64_t constant */
-#define LL(N)		N##LL
+#define LL(N)		(((int64_t)N##LL)
 
 /** Statistics counter type.  Not much excuse for not using a 64-bit
  * integer here, but if you're desperate and only run short
